changed id type to Long in the entire project

+++ backend/backend_with_jaxrs/src/main/java/com/example/backend_with_jaxrs/controllers/AppointmentController.java
@@ -30,7 +30,7 @@ public class AppointmentController {
     @Path("{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateAppointment(@PathParam("id") int id, Appointment appointment) throws CustomException {
+    public Response updateAppointment(@PathParam("id") Long id, Appointment appointment) throws CustomException {
         appointment.setId(id);
         Appointment updatedAppointment = AppointmentService.getInstance().updateAppointment(appointment);
         return Response.ok().entity(updatedAppointment).build();
