+++ backend/backend_with_jaxrs/src/main/java/com/example/backend_with_jaxrs/models/Client.java
@@ -12,7 +12,7 @@ public class Client extends User {
 
     }
 
-    public Client(int id) {
+    public Client(Long id) {
         super(id);
     }
